check for rogue riak_core_ring_manager and kill it
In the case where running eunit tests from the root dir of riak there
seems to be some sort of rogue riak_core_ring_manager process out
there.  I assume this is started by one of the unit tests in riak-core
or riak-kv but I didn't have time to hunt it down.  This whole
test_helper still seems a bit hackish but I just wanted to try and get
something working for now.